add more assertions
