+++ src/main/scala/com/github/agourlay/cornichon/util/Instances.scala
@@ -14,6 +14,11 @@ trait Instances extends AllInstances {
       fa.toIterator.map(_.show).mkString("Seq(", ", ", ")")
   }
 
+  implicit def showTraversable[A: Show]: Show[Traversable[A]] = new Show[Traversable[A]] {
+    def show(fa: Traversable[A]): String =
+      fa.toIterator.map(_.show).mkString("(", ", ", ")")
+  }
+
   implicit def showIndexedSeq[A: Show]: Show[IndexedSeq[A]] = new Show[IndexedSeq[A]] {
     def show(fa: IndexedSeq[A]): String =
       fa.toIterator.map(_.show).mkString("IndexedSeq(", ", ", ")")
